chore: amélioration de la gestion du refresh cote admin
refactorisation du code dans un inMemoryJWT manager
suppression de l'utilisation du global
